Eject assistant-event-hub from the core cycle
The generic pub/sub hub embedded guardian business logic: broadcastMessage
fire-and-forgot createCanonicalRequestForConfirmation, which dynamically
imported daemon/conversation-registry and confirmation-request-guardian-bridge.
Those two edges were the only thing keeping runtime/assistant-event-hub.ts
(in-degree 113, present in 87 of 191 circular chains) and its downstream in
the daemon-core SCC.

Move the promotion to the emitters of the confirmation_request event —
PermissionPrompter (interactive tool approvals) and the ACP route approval
gate — in a new permissions/confirmation-canonical-request.ts. The heavy
deps stay lazily imported (as they were in the hub) so the emitters' import
graph stays cheap and free of import-time DB access. The hub is now a pure
pub/sub primitive with no dependency on the registry or guardian bridge.

Core SCC: 327 -> 284 files (-43). Circular chains: 191 -> 140; none involve
the event hub anymore (was 87).
