+++ assistant/src/permissions/prompter.ts
@@ -7,6 +7,7 @@ import { redactSensitiveFields } from "../security/redaction.js";
 import type { ExecutionTarget } from "../tools/tool-types.js";
 import { AssistantError, ErrorCode } from "../util/errors.js";
 import { getLogger } from "../util/logger.js";
+import { createCanonicalRequestForConfirmation } from "./confirmation-canonical-request.js";
 import type { AllowlistOption, ScopeOption, UserDecision } from "./types.js";
 
 const log = getLogger("permission-prompter");
@@ -142,7 +143,9 @@ export class PermissionPrompter {
         signal.addEventListener("abort", onAbort, { once: true });
       }
 
-      this.sendToClient({
+      const confirmationMsg: ServerMessage & {
+        type: "confirmation_request";
+      } = {
         type: "confirmation_request",
         requestId,
         toolName,
@@ -170,7 +173,17 @@ export class PermissionPrompter {
         executionTarget,
         persistentDecisionsAllowed: persistentDecisionsAllowed ?? true,
         toolUseId,
-      });
+      };
+      this.sendToClient(confirmationMsg);
+
+      // Promote the confirmation to a canonical guardian request so channel
+      // guardian decisions (reactions, buttons, text) can resolve it.
+      if (conversationId) {
+        void createCanonicalRequestForConfirmation(
+          confirmationMsg,
+          conversationId,
+        );
+      }
 
       this.onStateChanged?.(requestId, "pending", "system", toolUseId);
     });
